Some changes required to compile on NT
git-svn-id: http://svn.cern.ch/guest/lhcb/Gauss/trunk@90221 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 41739644b06289120bd0cf3c394a2d9858c2bc65

+++ Gauss/Sim/GiGa/src/GiGaSensDet.cpp
@@ -72,8 +72,8 @@ StatusCode GiGaSensDet::initialize()
 	G4VSensitiveDetector::SensitiveDetectorName.remove(0,pos+1)    ; /// ATTENTION !!!
 	G4VSensitiveDetector::thePathName           = tmp              ; /// ATTENTION !!! 
 	G4VSensitiveDetector::thePathName.remove(pos+1,tmp.length()-1) ; /// ATTENTION !!! 
-	if( '/' != G4VSensitiveDetector::thePathName[0] ) 
-	  { G4VSensitiveDetector::thePathName.insert(0,'/'); }           /// ATTENTION !!! 
+	if( '/' != G4VSensitiveDetector::thePathName[(unsigned int)(0)] ) 
+  { G4VSensitiveDetector::thePathName.insert(0,'/'); }            /// ATTENTION !!! 
       } 
     ///
     G4VSensitiveDetector::fullPathName =                                  /// ATTENTION !!!
